backend/requirements.txt
numpy==1.17.2
scipy
setuptools
wheel
alembic==1.1.0
aniso8601==7.0.0
attrs==19.1.0
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
cycler==0.10.0
Cython==0.28.5
decorator==4.4.1
django-environ==0.4.5
dtw==1.4.0
elasticsearch==7.0.4
elasticsearch-dsl==7.0.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-Migrate==2.5.2
Flask-RESTful==0.3.5
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
imageio==2.6.1
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.14.0
jsonschema==3.0.2
kiwisolver==1.1.0
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
matrixprofile-ts==0.0.9
networkx==2.4
pandas==0.25.1
Pillow==6.2.1
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
PyWavelets==1.1.1
PyYAML==5.1.2
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.3
virtualenv==16.0.0
Werkzeug==0.15.5
Flask-WTF
pytest